+++ assets/js/app.js
@@ -20,4 +20,3 @@ import "phoenix_html"
 
 import socket from "./sockets/socket"
 import "./components/new_task"
-import "./components/new_user"